Queue bug fixed
BUG: once the client's lines were over, if player pressed Mouse 0 or space, the client would bug, randomizing its profile and reseting the enter animation

FIX: The ClientManager's CallNextClient function was being called by the player on DialogueManger's WaitToLeave coroutine without the canCallNextClient being checked with it was true. So, to fix it, added a check in the CallNextClient function to verify if the bool is true. If it is, the code runs as it should. If not, then return.